packages/api/pyproject.toml
[project]
name = "pixl-api"
version = "10.2.0"
description = "pixl HTTP API — thin FastAPI layer over pixl-cli"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pixl-cli",
    "pixl-engine[api]",
    "scalar-fastapi>=1.0.0",
    "httpx>=0.28.0",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.28.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
